requirements.txt
pandas
selenium
chromedriver-py
google-cloud-storage
webdriver-manager